# docs/api/openapi.json
{
  "components" : {
    "responses" : {
      "BadRequest" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "The request is malformed or fails validation; `fields` names the offending fields, never their values."
      },
      "Conflict" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "A conflict: duplicate, last administrator, or a concurrent change that survived the retries."
      },
      "Forbidden" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "The caller's role does not allow this (decided by the use case)."
      },
      "NotFound" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "Not found - also what an id of another association gets."
      },
      "RuleViolation" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "A business rule refuses it; `message` says which."
      },
      "TooManyRequests" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "Too many requests; see `Retry-After`."
      },
      "Unauthenticated" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "No valid access token (missing, expired, forged, or the account is no longer active)."
      }
    },
    "schemas" : {
...
      "ApiError" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "fields" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          }
        }
      },
...
      "AttendanceEntryRequest" : {
        "properties" : {
          "bookingId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "mark" : {
            "enum" : [ "ATTENDED", "NO_SHOW" ],
            "type" : "string"
          }
        },
        "required" : [ "bookingId", "mark" ],
        "type" : "object"
      },
      "AttendanceMarkedResponse" : {
        "properties" : {
          "session" : {
            "$ref" : "#/components/schemas/SessionResponse"
          },
          "warnedCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "BookableSessionResponse" : {
        "properties" : {
          "bookingOpensAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "capacity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "endsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "freeSeats" : {
            "format" : "int32",
            "type" : "integer"
          },
          "groupName" : {
            "type" : "string"
          },
          "myBookingStatus" : {
            "type" : "string"
          },
          "sessionId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "startsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "trainingGroupId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "waitlistSize" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "BookingCancelledResponse" : {
        "properties" : {
          "bookingId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "creditRefunded" : {
            "type" : "boolean"
          },
          "late" : {
            "type" : "boolean"
          },
          "promotedCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sessionId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "BookingPlacedResponse" : {
        "properties" : {
          "bookingId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "sessionId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "waitlistPosition" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "BookingView" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "memberId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "CancelSessionRequest" : {
        "properties" : {
          "reason" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "CapacityChangedResponse" : {
        "properties" : {
          "promotedCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "session" : {
            "$ref" : "#/components/schemas/SessionResponse"
          }
        },
        "type" : "object"
      },
      "ChangeCapacityRequest" : {
        "properties" : {
          "capacity" : {
            "format" : "int32",
            "maximum" : 1000,
            "minimum" : 1,
            "type" : "integer"
          }
        },
        "required" : [ "capacity" ],
        "type" : "object"
      },
...
      "MarkAttendanceRequest" : {
        "properties" : {
          "entries" : {
            "items" : {
              "$ref" : "#/components/schemas/AttendanceEntryRequest"
            },
            "maxItems" : 200,
            "minItems" : 1,
            "type" : "array"
          }
        },
        "required" : [ "entries" ],
        "type" : "object"
      },
...
      "SessionCancelledResponse" : {
        "properties" : {
          "bookingsCancelled" : {
            "format" : "int32",
            "type" : "integer"
          },
          "creditsRefunded" : {
            "format" : "int32",
            "type" : "integer"
          },
          "session" : {
            "$ref" : "#/components/schemas/SessionResponse"
          }
        },
        "type" : "object"
      },
      "SessionResponse" : {
        "properties" : {
          "bookings" : {
            "items" : {
              "$ref" : "#/components/schemas/BookingView"
            },
            "type" : "array"
          },
          "cancellationReason" : {
            "type" : "string"
          },
          "capacity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "coachId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "confirmedCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "endsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "freeSeats" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "startsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "trainingGroupId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "waitlistSize" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "The short-lived access token of POST /auth/login or /auth/refresh.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Session-by-session booking for amateur indoor volleyball associations. All paths are under /api/v1. The tenant\n(association) and the caller always come from the bearer token, never from a path, query or body; an id that\nbelongs to another association answers exactly like an id that does not exist (404). Every failure has the body\ndescribed by ApiError. Times are ISO-8601 instants (UTC); weekly schedules and dates are Europe/Lisbon local time.\nMoney is in cents.",
    "title" : "RegiVolley API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/sessions" : {
      "get" : {
        "operationId" : "sessionWeek",
        "parameters" : [ {
          "in" : "query",
          "name" : "weekOf",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/BookableSessionResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "session-controller" ]
      }
    },
    "/api/v1/sessions/{sessionId}/attendance" : {
      "put" : {
        "operationId" : "sessionAttendance",
        "parameters" : [ {
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MarkAttendanceRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AttendanceMarkedResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "session-controller" ]
      }
    },
    "/api/v1/sessions/{sessionId}/bookings" : {
      "post" : {
        "operationId" : "sessionBook",
        "parameters" : [ {
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BookingPlacedResponse"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "session-controller" ]
      }
    },
    "/api/v1/sessions/{sessionId}/bookings/{bookingId}/cancellation" : {
      "post" : {
        "operationId" : "sessionCancelBooking",
        "parameters" : [ {
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "bookingId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BookingCancelledResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "session-controller" ]
      }
    },
    "/api/v1/sessions/{sessionId}/cancellation" : {
      "post" : {
        "operationId" : "sessionCancel",
        "parameters" : [ {
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CancelSessionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionCancelledResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "session-controller" ]
      }
    },
    "/api/v1/sessions/{sessionId}/capacity" : {
      "put" : {
        "operationId" : "sessionCapacity",
        "parameters" : [ {
          "in" : "path",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangeCapacityRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CapacityChangedResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthenticated"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "422" : {
            "$ref" : "#/components/responses/RuleViolation"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "session-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}